Simplify example.
